[Gold V] Title: 입국심사, Time: 412 ms, Memory: 27668 KB -BaekjoonHub

+++ 백준/Gold/3079. 입국심사/입국심사.java
@@ -1,11 +1,19 @@
 import java.io.*;
 import java.util.Arrays;
 
+/** 3017. 입국심사
+ * 메모리
+ * 시간
+ *
+ * 이분탐색
+ * getCount 할 때 count가 personN을 넘으면 끊어줘야 한다. -> 안 하면 32%에서 틀림
+ * personN이 1억까지기 나올 수 있기 때문에 end값이 int 범위를 넘을 수 있다.
+ */
 public class Main {
 
 	static BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 	static int N;
-	static long personN;
+	static int personN;
 	static long[] Times;
 
 	static long result;
@@ -52,7 +60,7 @@ public class Main {
 	private static void init() throws Exception {
 		String[] inputs = br.readLine().split(" ");
 		N = Integer.parseInt(inputs[0]);
-		personN = Long.parseLong(inputs[1]);
+		personN = Integer.parseInt(inputs[1]);
 
 		Times = new long[(int)N];
 		for (int i = 0; i < N; i++) {
